pyproject.toml
[project]
name = "photon-platform-publish"
dynamic = ["version"]
authors = [
  { name="phi ARCHITECT", email="github@phiarchitect.com" },
]
description = "all the tools and themes for publishing with Sphinx"
readme = "README.rst"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "sphinx",
    "Pillow",
    "jinja2",
    "myst_parser",
    "sphinx-carousel",
    "sphinxcontrib-jquery",
    "sphinxcontrib-youtube",
    "sphinx_revealjs",
    "sphinxext-opengraph",
    "libsass",
    "click",
    "sphinx-autoapi",
    "matplotlib",
    "graphviz",

    #  "sphinxcontrib.bibtex",
]

[project.scripts]
publish = "photon_platform.publish.__main__:run"

[project.entry-points."sphinx.html_themes"]
photon = "photon_platform.publish.themes.photon"
foundation = "photon_platform.publish.themes.foundation"


[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["photon_platform"]
namespaces = true

[tool.setuptools.dynamic]
version = {attr = "photon_platform.publish.__version__"}

[project.urls]
"Homepage" = "https://photon-platform.github.io/publish"
"Website" = "https://photon-platform.github.io/publish"
"Repository" = "https://github.com/photon-platform/publish"
"Issues" = "https://github.com/photon-platform/publish/issues"
